requirements.txt
inference-sdk>=1.0.0
Pillow>=11.0.0
